ssh-util: remove debug_name
A forthcoming commit will introduce an SSH tunnel manager that will
allow the same tunnel to be shared between multiple contexts. So instead
of having the caller pass in a debug name, use the remote host/port as
the debug name directly (i.e., in the name of the Tokio task that is
servicing the tunnel).