sql[-parser]: rearrange COPY types
Rejigger the types underlying the parsing and planning of the COPY
statement to make it impossible to represent some invalid COPY
statements. For example, it is no longer possible to represent `COPY
(SELECT 1) FROM STDIN`.

This is purely refactoring and has no behavior changes, besides some
minor changes to error messages.